+++ tests/testthat/test_space_filling.R
@@ -3,6 +3,8 @@ context("space filling designs")
 
 # ------------------------------------------------------------------------------
 
+source("helper-functions.R")
+
 # ------------------------------------------------------------------------------
 
 test_that('max entropy designs', {
@@ -100,3 +102,21 @@ test_that('latin square designs', {
 })
 
 
+test_that('grid attributes', {
+  p <- parameters(penalty(), mixture())
+  expect_true(proper_grid(grid_latin_hypercube(p), "grid_latin_hypercube"))
+  expect_true(proper_grid(
+    grid_latin_hypercube(penalty(), mixture()),
+    "grid_latin_hypercube"
+  ))
+  expect_true(proper_grid(grid_latin_hypercube(list(
+    penalty(), mixture()
+  )), "grid_latin_hypercube"))
+
+  expect_true(proper_grid(grid_max_entropy(p), "grid_max_entropy"))
+  expect_true(proper_grid(grid_max_entropy(penalty(), mixture()), "grid_max_entropy"))
+  expect_true(proper_grid(grid_max_entropy(list(
+    penalty(), mixture()
+  )), "grid_max_entropy"))
+
+})